% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manager.R
\name{mgr_get_config}
\alias{mgr_get_config}
\title{Retrieve the configuration of a method.}
\usage{
mgr_get_config(method)
}
\arguments{
\item{method}{A representation method.}
}
\value{
A \code{list} of configuration parameters of \code{method}.
}
\description{
Retrieve the configuration of a method.
}
