% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manager.R
\name{mgr_init}
\alias{mgr_init}
\title{Initialize a representation.}
\usage{
mgr_init(method_name)
}
\arguments{
\item{method_name}{A method name.}
}
\value{
A representation method.
}
\description{
Initialize a representation.
}
