% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manager.R
\name{mgr_set_config}
\alias{mgr_set_config}
\title{Set the configuration of a method.}
\usage{
mgr_set_config(method, config)
}
\arguments{
\item{method}{A representation method.}

\item{config}{A \code{list} of configuration parameters for \code{method}.}
}
\value{
The \code{method} with updated parameters.
}
\description{
Set the configuration of a method.
}
